+++ eggs/Naaya/Products/NaayaCore/AuthenticationTool/plugins/ldap_cache.py
@@ -66,7 +66,15 @@ class Cache(object):
         Retrieve a record from this cache. Raises KeyError if record not found.
         """
         if self.timestamp is None:
-            self.update()
+            try:
+                self.update()
+            except Exception, e:
+                log.exception("Update failed for empty cache")
+                # behave as cache miss
+                if default is _raise_if_not_found:
+                    raise
+                else:
+                    return default
 
         try:
             record = self.users[user_dn]