+++ eggs/Naaya/CHANGELOG.rst
@@ -1,5 +1,6 @@
 3.2.14 (unreleased)
 -------------------
+* ldap_cache: unsuccessful init update behaves as cache miss [simiamih]
 
 3.2.13 (2013-01-11)
 -------------------